ANALYSIS OF JUL 11, 2022
The random forest models have not changed, but I made significant adjustments to the maps as I did determine that it was taking the last entry for each state to make each respective map. Instead, I elected to make (probably inefficiently) many filtering functions that take either the average or mode of the values for each state and then show on a map. I would like to discuss this more tomorrow to determine what is appropriate (between the central tendency measures) and if it is even relevant/needed.