CLEANED RMD FILE, ONE TREE PULLED, PARTIAL PLOTS
In addition to the changes graciously made by Sam today, I have gone ahead and written some code for the partial plots of each variable, emphasizing the risk of >95% percentile BMI. Curious trends are discussed in the email, but this code is left here for now. Partial plots begin on line 213.