US MAP PLOTS
This file includes various U.S. maps detailing the ranges of variable values between states. One concern I had last week was that the function plot_usmap() was only taking the last entry of each state to fill in, and this fear was confirmed when I instead made new dataframes that either attached the mean or mode statistic to each state. I would have posted this in the results folder, but for some reason upon knitting, it does not seem to recognize that I have installed and loaded the RColorBrewer package (for colors.)